+++ CotEditor/Sources/EditorTextView+Scaling.swift
@@ -9,7 +9,7 @@
  
  ------------------------------------------------------------------------------
  
- © 2016 1024jp
+ © 2016-2018 1024jp
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -56,7 +56,7 @@ extension EditorTextView {
         }
         
         // sanitize final scale
-        if event.phase.contains(.ended) && fabs(scale - 1.0) < 0.05 {
+        if event.phase.contains(.ended), fabs(scale - 1.0) < 0.05 {
             scale = 1.0
         }