+++ CotEditor/Sources/ShortcutKeyField.swift
@@ -10,7 +10,7 @@
  ------------------------------------------------------------------------------
  
  © 2004-2007 nakamuxu
- © 2014-2016 1024jp
+ © 2014-2018 1024jp
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -46,8 +46,10 @@ final class ShortcutKeyField: NSTextField {
         
         self.keyDownMonitor = NSEvent.addLocalMonitorForEvents(matching: .keyDown, handler: { [weak self] (event: NSEvent) -> NSEvent? in
             
-            guard var charsIgnoringModifiers = event.charactersIgnoringModifiers else { return event }
-            guard !charsIgnoringModifiers.isEmpty else { return event }
+            guard
+                var charsIgnoringModifiers = event.charactersIgnoringModifiers,
+                !charsIgnoringModifiers.isEmpty
+                else { return event }
             
             var modifierMask = event.modifierFlags
             
@@ -64,7 +66,7 @@ final class ShortcutKeyField: NSTextField {
             
             // remove unwanted Shift
             let ignoringShiftSet = CharacterSet(charactersIn: "`~!@#$%^&()_{}|\":<>?=/*-+.'")
-            if ignoringShiftSet.contains(charsIgnoringModifiers.unicodeScalars.first!) && modifierMask.contains(.shift) {
+            if ignoringShiftSet.contains(charsIgnoringModifiers.unicodeScalars.first!), modifierMask.contains(.shift) {
                 modifierMask.remove(.shift)
             }