+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 Change Log
 ==========================
 
-develop
+3.3.0 (224)
+--------------------------
+
+### Fixes
+
+- [beta] Fix an issue where iCloud document storage was not created.
+- [beta] Fix an issue where find result message in the find panel fields was not shown.
+
+
+
+3.3.0-beta.3 (220)
 --------------------------
 
 ### Improvements